requirements.txt
Cython==0.24
Pillow==3.2.0
pytesseract==0.1.6
tesseract-ocr==0.0.1
